improvements

+++ entry.sh
@@ -7,13 +7,21 @@ fi
 output_path=$(resume_converter "$@" | tail -n 1)
 
 if [ -f "$output_path" ]; then
+    dir=$(dirname "$output_path")
+    file=$(basename "$output_path")
+    
+    if [ "$dir" != "." ]; then
+        cd "$dir" || { echo "Failed to change directory to $dir"; exit 1; }
+        output_path="$file"
+    fi
+
     "$LATEX_COMPILER" "$output_path"
     if [ $? -eq 0 ]; then
         echo "The file '$output_path' was successfully compiled."
 
         output_path_no_ext="${output_path%.tex}"
-        extesions_to_remove=("aux" "log" "out" "synctex.gz")
-        for ext in "${extesions_to_remove[@]}"; do
+        extensions_to_remove=("aux" "log" "out" "synctex.gz")
+        for ext in "${extensions_to_remove[@]}"; do
             if [ -f "$output_path_no_ext.$ext" ]; then
                 rm "$output_path_no_ext.$ext"
                 echo "The file '$output_path_no_ext.$ext' was removed."
